// server/src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_CONFIRMED = 'ORDER_CONFIRMED',
  ORDER_SHIPPED = 'ORDER_SHIPPED',
  OUT_FOR_DELIVERY = 'OUT_FOR_DELIVERY',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  REFUND_INITIATED = 'REFUND_INITIATED',
  REFUND_COMPLETED = 'REFUND_COMPLETED',
  RETURN_APPROVED = 'RETURN_APPROVED',
  COUPON_AVAILABLE = 'COUPON_AVAILABLE',
  WISHLIST_BACK_IN_STOCK = 'WISHLIST_BACK_IN_STOCK',
  WISHLIST_PRICE_DROP = 'WISHLIST_PRICE_DROP',
  NEW_ORDER = 'NEW_ORDER',
  NEW_REVIEW = 'NEW_REVIEW',
  LOW_STOCK = 'LOW_STOCK',
  OUT_OF_STOCK = 'OUT_OF_STOCK',
  NEW_MERCHANT_REGISTERED = 'NEW_MERCHANT_REGISTERED',
  PRODUCT_REPORTED = 'PRODUCT_REPORTED',
}