// server/src/email/enums/email-template.enum.ts
export enum EmailTemplate {
  ORDER_CREATED = 'order-created',
  ORDER_CONFIRMED = 'order-confirmed',
  ORDER_SHIPPED = 'order-shipped',
  ORDER_DELIVERED = 'order-delivered',
  RETURN_APPROVED = 'return-approved',
  REFUND_COMPLETED = 'refund-completed',
  PASSWORD_CHANGED = 'password-changed',
  MERCHANT_REGISTRATION = 'merchant-registration',
}